Create find_prime_numbers_eratostheneses.py
Using the Sieve of Eratostheneses we can find prime number in a bounded range, i.e, prime number less than a given number.

I think we can't (I could be wrong) find the nth prime number with it even if we keep incrementing the guessed upper bound till we obtain the nth prime number. So for example in the code below, the 10th prime number is 29 and the 62nd prime number is 293; it's difficult to have an accurate initial guess of the upper bound. If we have a larger than necessary upper bound, our O(n) will be much large for no reason. On the other hand, if we keep incrementing the upper bound as we go in the algorithm, then we will have to write an inefficient code to eliminate all new numbers added that are divisible by already found prime numbers.